[test] Catch error between tests
The issue was found in https://github.com/mui-org/material-ui-x/pull/543.
I also create a new setupKarma.js file so the X repo can import it directly.
